detector/requirements.txt
flask==3.0.3
requests==2.32.3
pyyaml==6.0.2
psutil==6.0.0
